abstract backend resolvers
