+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/main/java/org/hypertrace/traceenricher/enrichment/enrichers/resolver/backend/CassandraBackendResolver.java
@@ -1,61 +1,46 @@
 package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
 
-import java.util.HashMap;
+import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
-import org.apache.commons.lang3.StringUtils;
-import org.hypertrace.core.datamodel.AttributeValue;
 import org.hypertrace.core.datamodel.Event;
 import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
-import org.hypertrace.core.datamodel.shared.trace.AttributeValueCreator;
-import org.hypertrace.entity.data.service.v1.Entity.Builder;
 import org.hypertrace.semantic.convention.utils.db.DbSemanticConventionUtils;
-import org.hypertrace.traceenricher.enrichedspan.constants.EnrichedSpanConstants;
-import org.hypertrace.traceenricher.enrichedspan.constants.v1.Backend;
 import org.hypertrace.traceenricher.enrichment.enrichers.BackendType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class CassandraBackendResolver extends AbstractBackendResolver {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(CassandraBackendResolver.class);
-  private static final String BACKEND_OPERATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_OPERATION);
-  private static final String BACKEND_DESTINATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_DESTINATION);
-
   public CassandraBackendResolver(FqnResolver fqnResolver) {
     super(fqnResolver);
   }
 
   @Override
-  public Optional<BackendInfo> resolve(Event event, StructuredTraceGraph structuredTraceGraph) {
-    if (!DbSemanticConventionUtils.isCassandraBackend(event)) {
-      return Optional.empty();
-    }
+  public boolean isValidBackend(Event event) {
+    return DbSemanticConventionUtils.isCassandraBackend(event);
+  }
 
-    Optional<String> backendURI = DbSemanticConventionUtils.getCassandraURI(event);
+  @Override
+  public BackendType getBackendType(Event event) {
+    return BackendType.CASSANDRA;
+  }
 
-    if (backendURI.isEmpty() || StringUtils.isEmpty(backendURI.get())) {
-      LOGGER.warn("Unable to infer a cassandra backend from event: {}", event);
-      return Optional.empty();
-    }
+  @Override
+  public Optional<String> getBackendUri(Event event, StructuredTraceGraph structuredTraceGraph) {
+    return DbSemanticConventionUtils.getCassandraURI(event);
+  }
 
-    final Builder entityBuilder =
-        getBackendEntityBuilder(BackendType.CASSANDRA, backendURI.get(), event);
+  @Override
+  public Map<String, org.hypertrace.entity.data.service.v1.AttributeValue> getEntityAttributes(
+      Event event) {
+    return Collections.emptyMap();
+  }
 
-    Map<String, AttributeValue> enrichedAttributes = new HashMap<>();
-    Optional<String> cassandraOperation = DbSemanticConventionUtils.getDbOperation(event);
-    cassandraOperation.ifPresent(
-        operation ->
-            enrichedAttributes.put(
-                BACKEND_OPERATION_ATTR, AttributeValueCreator.create(operation)));
-    Optional<String> cassandraDestination =
-        DbSemanticConventionUtils.getDestinationForCassandra(event);
-    cassandraDestination.ifPresent(
-        destination ->
-            enrichedAttributes.put(
-                BACKEND_DESTINATION_ATTR, AttributeValueCreator.create(destination)));
-    return Optional.of(new BackendInfo(entityBuilder.build(), enrichedAttributes));
+  @Override
+  public Optional<String> getBackendOperation(Event event) {
+    return DbSemanticConventionUtils.getDbOperation(event);
+  }
+
+  @Override
+  public Optional<String> getBackendDestination(Event event) {
+    return DbSemanticConventionUtils.getDestinationForCassandra(event);
   }
 }
