+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/main/java/org/hypertrace/traceenricher/enrichment/enrichers/resolver/backend/ElasticSearchBackendResolver.java
@@ -1,64 +1,50 @@
 package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
 
-import static org.hypertrace.traceenricher.util.EnricherUtil.setAttributeIfExist;
+import static org.hypertrace.traceenricher.util.EnricherUtil.getAttributesForFirstExistingKey;
 
-import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import org.apache.commons.lang3.StringUtils;
-import org.hypertrace.core.datamodel.AttributeValue;
 import org.hypertrace.core.datamodel.Event;
 import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
-import org.hypertrace.core.datamodel.shared.trace.AttributeValueCreator;
 import org.hypertrace.core.semantic.convention.constants.db.OTelDbSemanticConventions;
-import org.hypertrace.entity.data.service.v1.Entity.Builder;
 import org.hypertrace.semantic.convention.utils.db.DbSemanticConventionUtils;
-import org.hypertrace.traceenricher.enrichedspan.constants.EnrichedSpanConstants;
-import org.hypertrace.traceenricher.enrichedspan.constants.v1.Backend;
 import org.hypertrace.traceenricher.enrichment.enrichers.BackendType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class ElasticSearchBackendResolver extends AbstractBackendResolver {
-  private static final Logger LOGGER = LoggerFactory.getLogger(ElasticSearchBackendResolver.class);
-  private static final String BACKEND_OPERATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_OPERATION);
-  private static final String BACKEND_DESTINATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_DESTINATION);
-
   public ElasticSearchBackendResolver(FqnResolver fqnResolver) {
     super(fqnResolver);
   }
 
   @Override
-  public Optional<BackendInfo> resolve(Event event, StructuredTraceGraph structuredTraceGraph) {
-    if (!DbSemanticConventionUtils.isElasticSearchBackend(event)) {
-      return Optional.empty();
-    }
+  public boolean isValidBackend(Event event) {
+    return DbSemanticConventionUtils.isElasticSearchBackend(event);
+  }
 
-    Optional<String> backendURI = DbSemanticConventionUtils.getElasticSearchURI(event);
+  @Override
+  public BackendType getBackendType(Event event) {
+    return BackendType.ELASTICSEARCH;
+  }
 
-    if (backendURI.isEmpty() || StringUtils.isEmpty(backendURI.get())) {
-      LOGGER.warn("Unable to infer a elasticsearch backend from event: {}", event);
-      return Optional.empty();
-    }
+  @Override
+  public Optional<String> getBackendUri(Event event, StructuredTraceGraph structuredTraceGraph) {
+    return DbSemanticConventionUtils.getElasticSearchURI(event);
+  }
 
-    final Builder entityBuilder =
-        getBackendEntityBuilder(BackendType.ELASTICSEARCH, backendURI.get(), event);
-    setAttributeIfExist(event, entityBuilder, OTelDbSemanticConventions.DB_OPERATION.getValue());
+  @Override
+  public Map<String, org.hypertrace.entity.data.service.v1.AttributeValue> getEntityAttributes(
+      Event event) {
+    return getAttributesForFirstExistingKey(
+        event, List.of(OTelDbSemanticConventions.DB_OPERATION.getValue()));
+  }
 
-    Map<String, AttributeValue> enrichedAttributes = new HashMap<>();
-    Optional<String> esOperation = DbSemanticConventionUtils.getDbOperationForElasticSearch(event);
-    esOperation.ifPresent(
-        operation ->
-            enrichedAttributes.put(
-                BACKEND_OPERATION_ATTR, AttributeValueCreator.create(operation)));
-    Optional<String> esDestination =
-        DbSemanticConventionUtils.getDestinationForElasticsearch(event);
-    esDestination.ifPresent(
-        destination ->
-            enrichedAttributes.put(
-                BACKEND_DESTINATION_ATTR, AttributeValueCreator.create(destination)));
-    return Optional.of(new BackendInfo(entityBuilder.build(), enrichedAttributes));
+  @Override
+  public Optional<String> getBackendOperation(Event event) {
+    return DbSemanticConventionUtils.getDbOperationForElasticSearch(event);
+  }
+
+  @Override
+  public Optional<String> getBackendDestination(Event event) {
+    return DbSemanticConventionUtils.getDestinationForElasticsearch(event);
   }
 }
