+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/main/java/org/hypertrace/traceenricher/enrichment/enrichers/resolver/backend/MongoBackendResolver.java
@@ -1,69 +1,62 @@
 package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
 
-import static org.hypertrace.traceenricher.util.EnricherUtil.setAttributeForFirstExistingKey;
-import static org.hypertrace.traceenricher.util.EnricherUtil.setAttributeIfExist;
+import static org.hypertrace.traceenricher.util.EnricherUtil.getAttributesForFirstExistingKey;
 
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import org.apache.commons.lang3.StringUtils;
-import org.hypertrace.core.datamodel.AttributeValue;
 import org.hypertrace.core.datamodel.Event;
 import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
-import org.hypertrace.core.datamodel.shared.trace.AttributeValueCreator;
-import org.hypertrace.entity.data.service.v1.Entity.Builder;
 import org.hypertrace.semantic.convention.utils.db.DbSemanticConventionUtils;
-import org.hypertrace.traceenricher.enrichedspan.constants.EnrichedSpanConstants;
-import org.hypertrace.traceenricher.enrichedspan.constants.v1.Backend;
 import org.hypertrace.traceenricher.enrichment.enrichers.BackendType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class MongoBackendResolver extends AbstractBackendResolver {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(MongoBackendResolver.class);
   private static final String RAW_MONGO_NAMESPACE = "NAMESPACE";
-  private static final String BACKEND_OPERATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_OPERATION);
-  private static final String BACKEND_DESTINATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_DESTINATION);
 
   public MongoBackendResolver(FqnResolver fqnResolver) {
     super(fqnResolver);
   }
 
   @Override
-  public Optional<BackendInfo> resolve(Event event, StructuredTraceGraph structuredTraceGraph) {
-    if (!DbSemanticConventionUtils.isMongoBackend(event)) {
-      return Optional.empty();
-    }
+  public boolean isValidBackend(Event event) {
+    return DbSemanticConventionUtils.isMongoBackend(event);
+  }
 
-    Optional<String> backendURI = DbSemanticConventionUtils.getMongoURI(event);
+  @Override
+  public BackendType getBackendType(Event event) {
+    return BackendType.MONGO;
+  }
 
-    if (backendURI.isEmpty() || StringUtils.isEmpty(backendURI.get())) {
-      LOGGER.warn("Unable to infer a mongo backend from event: {}", event);
-      return Optional.empty();
-    }
+  @Override
+  public Optional<String> getBackendUri(Event event, StructuredTraceGraph structuredTraceGraph) {
+    return DbSemanticConventionUtils.getMongoURI(event);
+  }
 
-    final Builder entityBuilder =
-        getBackendEntityBuilder(BackendType.MONGO, backendURI.get(), event);
-    setAttributeIfExist(event, entityBuilder, RAW_MONGO_NAMESPACE);
-    setAttributeForFirstExistingKey(
-        event, entityBuilder, DbSemanticConventionUtils.getAttributeKeysForMongoNamespace());
-    setAttributeForFirstExistingKey(
-        event, entityBuilder, DbSemanticConventionUtils.getAttributeKeysForMongoOperation());
+  @Override
+  public Map<String, org.hypertrace.entity.data.service.v1.AttributeValue> getEntityAttributes(
+      Event event) {
+    Map<String, org.hypertrace.entity.data.service.v1.AttributeValue> entityAttributes =
+        new HashMap<>();
+    entityAttributes.putAll(getAttributesForFirstExistingKey(event, List.of(RAW_MONGO_NAMESPACE)));
+    entityAttributes.putAll(
+        getAttributesForFirstExistingKey(
+            event, DbSemanticConventionUtils.getAttributeKeysForMongoNamespace()));
+    entityAttributes.putAll(
+        getAttributesForFirstExistingKey(
+            event, DbSemanticConventionUtils.getAttributeKeysForMongoOperation()));
 
-    Map<String, AttributeValue> enrichedAttributes = new HashMap<>();
-    Optional<String> mongoOperation = DbSemanticConventionUtils.getDbOperationForMongo(event);
-    mongoOperation.ifPresent(
-        operation ->
-            enrichedAttributes.put(
-                BACKEND_OPERATION_ATTR, AttributeValueCreator.create(operation)));
-    Optional<String> mongoDestination = DbSemanticConventionUtils.getDestinationForMongo(event);
-    mongoDestination.ifPresent(
-        destination ->
-            enrichedAttributes.put(
-                BACKEND_DESTINATION_ATTR, AttributeValueCreator.create(destination)));
-    return Optional.of(new BackendInfo(entityBuilder.build(), enrichedAttributes));
+    return Collections.unmodifiableMap(entityAttributes);
+  }
+
+  @Override
+  public Optional<String> getBackendOperation(Event event) {
+    return DbSemanticConventionUtils.getDbOperationForMongo(event);
+  }
+
+  @Override
+  public Optional<String> getBackendDestination(Event event) {
+    return DbSemanticConventionUtils.getDestinationForMongo(event);
   }
 }
