+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/main/java/org/hypertrace/traceenricher/enrichment/enrichers/resolver/backend/RedisBackendResolver.java
@@ -1,73 +1,64 @@
 package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
 
-import static org.hypertrace.traceenricher.util.EnricherUtil.setAttributeIfExist;
+import static org.hypertrace.traceenricher.util.EnricherUtil.getAttributesForFirstExistingKey;
 
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import org.apache.commons.lang3.StringUtils;
-import org.hypertrace.core.datamodel.AttributeValue;
 import org.hypertrace.core.datamodel.Event;
 import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
-import org.hypertrace.core.datamodel.shared.trace.AttributeValueCreator;
 import org.hypertrace.core.semantic.convention.constants.db.OTelDbSemanticConventions;
 import org.hypertrace.core.span.constants.RawSpanConstants;
 import org.hypertrace.core.span.constants.v1.Redis;
-import org.hypertrace.entity.data.service.v1.Entity.Builder;
 import org.hypertrace.semantic.convention.utils.db.DbSemanticConventionUtils;
-import org.hypertrace.traceenricher.enrichedspan.constants.EnrichedSpanConstants;
-import org.hypertrace.traceenricher.enrichedspan.constants.v1.Backend;
 import org.hypertrace.traceenricher.enrichment.enrichers.BackendType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class RedisBackendResolver extends AbstractBackendResolver {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(RedisBackendResolver.class);
-  private static final String BACKEND_OPERATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_OPERATION);
-  private static final String BACKEND_DESTINATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_DESTINATION);
-
   public RedisBackendResolver(FqnResolver fqnResolver) {
     super(fqnResolver);
   }
 
   @Override
-  public Optional<BackendInfo> resolve(Event event, StructuredTraceGraph structuredTraceGraph) {
-    if (!DbSemanticConventionUtils.isRedisBackend(event)) {
-      return Optional.empty();
-    }
+  public boolean isValidBackend(Event event) {
+    return DbSemanticConventionUtils.isRedisBackend(event);
+  }
 
-    Optional<String> backendURI = DbSemanticConventionUtils.getRedisURI(event);
+  @Override
+  public BackendType getBackendType(Event event) {
+    return BackendType.REDIS;
+  }
 
-    if (backendURI.isEmpty()) {
-      return Optional.empty();
-    }
+  @Override
+  public Optional<String> getBackendUri(Event event, StructuredTraceGraph structuredTraceGraph) {
+    return DbSemanticConventionUtils.getRedisURI(event);
+  }
 
-    if (StringUtils.isEmpty(backendURI.get())) {
-      LOGGER.warn("Unable to infer a redis backend from event: {}", event);
-      return Optional.empty();
-    }
+  @Override
+  public Map<String, org.hypertrace.entity.data.service.v1.AttributeValue> getEntityAttributes(
+      Event event) {
+    Map<String, org.hypertrace.entity.data.service.v1.AttributeValue> entityAttributes =
+        new HashMap<>();
+    entityAttributes.putAll(
+        getAttributesForFirstExistingKey(
+            event, List.of(RawSpanConstants.getValue(Redis.REDIS_COMMAND))));
+    entityAttributes.putAll(
+        getAttributesForFirstExistingKey(
+            event, List.of(RawSpanConstants.getValue(Redis.REDIS_ARGS))));
+    entityAttributes.putAll(
+        getAttributesForFirstExistingKey(
+            event, List.of(OTelDbSemanticConventions.DB_CONNECTION_STRING.getValue())));
+    return Collections.unmodifiableMap(entityAttributes);
+  }
 
-    final Builder entityBuilder =
-        getBackendEntityBuilder(BackendType.REDIS, backendURI.get(), event);
-    setAttributeIfExist(event, entityBuilder, RawSpanConstants.getValue(Redis.REDIS_COMMAND));
-    setAttributeIfExist(event, entityBuilder, RawSpanConstants.getValue(Redis.REDIS_ARGS));
-    setAttributeIfExist(
-        event, entityBuilder, OTelDbSemanticConventions.DB_CONNECTION_STRING.getValue());
+  @Override
+  public Optional<String> getBackendOperation(Event event) {
+    return DbSemanticConventionUtils.getDbOperationForRedis(event);
+  }
 
-    Map<String, AttributeValue> enrichedAttributes = new HashMap<>();
-    Optional<String> redisOperation = DbSemanticConventionUtils.getDbOperationForRedis(event);
-    redisOperation.ifPresent(
-        operation ->
-            enrichedAttributes.put(
-                BACKEND_OPERATION_ATTR, AttributeValueCreator.create(operation)));
-    Optional<String> redisDestination = DbSemanticConventionUtils.getDestinationForRedis(event);
-    redisDestination.ifPresent(
-        destination ->
-            enrichedAttributes.put(
-                BACKEND_DESTINATION_ATTR, AttributeValueCreator.create(destination)));
-    return Optional.of(new BackendInfo(entityBuilder.build(), enrichedAttributes));
+  @Override
+  public Optional<String> getBackendDestination(Event event) {
+    return DbSemanticConventionUtils.getDestinationForRedis(event);
   }
 }
