+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/main/java/org/hypertrace/traceenricher/enrichment/enrichers/resolver/backend/SqsBackendResolver.java
@@ -1,66 +1,46 @@
 package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
 
-import java.util.HashMap;
+import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
-import org.apache.commons.lang3.StringUtils;
-import org.hypertrace.core.datamodel.AttributeValue;
 import org.hypertrace.core.datamodel.Event;
 import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
-import org.hypertrace.core.datamodel.shared.trace.AttributeValueCreator;
-import org.hypertrace.entity.data.service.v1.Entity;
 import org.hypertrace.semantic.convention.utils.messaging.MessagingSemanticConventionUtils;
-import org.hypertrace.traceenricher.enrichedspan.constants.EnrichedSpanConstants;
-import org.hypertrace.traceenricher.enrichedspan.constants.v1.Backend;
 import org.hypertrace.traceenricher.enrichment.enrichers.BackendType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class SqsBackendResolver extends AbstractBackendResolver {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(SqsBackendResolver.class);
-  private static final String BACKEND_OPERATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_OPERATION);
-  private static final String BACKEND_DESTINATION_ATTR =
-      EnrichedSpanConstants.getValue(Backend.BACKEND_DESTINATION);
-
   public SqsBackendResolver(FqnResolver fqnResolver) {
     super(fqnResolver);
   }
 
   @Override
-  public Optional<BackendInfo> resolve(Event event, StructuredTraceGraph structuredTraceGraph) {
-    if (!MessagingSemanticConventionUtils.isSqsBackend(event)) {
-      return Optional.empty();
-    }
+  public boolean isValidBackend(Event event) {
+    return MessagingSemanticConventionUtils.isSqsBackend(event);
+  }
 
-    Optional<String> backendURI = MessagingSemanticConventionUtils.getSqsBackendURI(event);
+  @Override
+  public BackendType getBackendType(Event event) {
+    return BackendType.SQS;
+  }
 
-    if (backendURI.isEmpty() || StringUtils.isEmpty(backendURI.get())) {
-      LOGGER.warn("Unable to infer a SQS backend from event: {}", event);
-      return Optional.empty();
-    }
-    /*
-    todo: should we clean protocol constants for rabbit_mq, mongo?
-    this is not added into enriched spans constants proto as there want be http.url
-    with sqs://xyz:2323.
-    * */
-    Entity.Builder entityBuilder =
-        getBackendEntityBuilder(BackendType.SQS, backendURI.get(), event);
+  @Override
+  public Optional<String> getBackendUri(Event event, StructuredTraceGraph structuredTraceGraph) {
+    return MessagingSemanticConventionUtils.getSqsBackendURI(event);
+  }
 
-    Map<String, AttributeValue> enrichedAttributes = new HashMap<>();
-    Optional<String> sqsOperation = MessagingSemanticConventionUtils.getMessagingOperation(event);
-    sqsOperation.ifPresent(
-        operation ->
-            enrichedAttributes.put(
-                BACKEND_OPERATION_ATTR, AttributeValueCreator.create(operation)));
-    Optional<String> sqsDestination =
-        MessagingSemanticConventionUtils.getMessagingDestination(event);
-    sqsDestination.ifPresent(
-        destination ->
-            enrichedAttributes.put(
-                BACKEND_DESTINATION_ATTR, AttributeValueCreator.create(destination)));
+  @Override
+  public Map<String, org.hypertrace.entity.data.service.v1.AttributeValue> getEntityAttributes(
+      Event event) {
+    return Collections.emptyMap();
+  }
 
-    return Optional.of(new BackendInfo(entityBuilder.build(), enrichedAttributes));
+  @Override
+  public Optional<String> getBackendOperation(Event event) {
+    return MessagingSemanticConventionUtils.getMessagingOperation(event);
+  }
+
+  @Override
+  public Optional<String> getBackendDestination(Event event) {
+    return MessagingSemanticConventionUtils.getMessagingDestination(event);
   }
 }
