+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/test/java/org/hypertrace/traceenricher/enrichment/enrichers/resolver/backend/MongoBackendResolverTest.java
@@ -1,0 +1,358 @@
+package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
+
+import static org.hypertrace.traceenricher.TestUtil.buildAttributeValue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.mock;
+
+import java.nio.ByteBuffer;
+import java.util.Arrays;
+import java.util.Map;
+import org.hypertrace.core.datamodel.AttributeValue;
+import org.hypertrace.core.datamodel.Attributes;
+import org.hypertrace.core.datamodel.Event;
+import org.hypertrace.core.datamodel.EventRef;
+import org.hypertrace.core.datamodel.EventRefType;
+import org.hypertrace.core.datamodel.MetricValue;
+import org.hypertrace.core.datamodel.Metrics;
+import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
+import org.hypertrace.core.datamodel.shared.trace.AttributeValueCreator;
+import org.hypertrace.core.semantic.convention.constants.db.OTelDbSemanticConventions;
+import org.hypertrace.core.semantic.convention.constants.span.OTelSpanSemanticConventions;
+import org.hypertrace.core.span.constants.v1.Mongo;
+import org.hypertrace.entity.constants.v1.BackendAttribute;
+import org.hypertrace.entity.data.service.v1.Entity;
+import org.hypertrace.traceenricher.enrichedspan.constants.v1.Backend;
+import org.hypertrace.traceenricher.util.Constants;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+public class MongoBackendResolverTest {
+  private static final String MONGO_URL = "mongo:27017";
+
+  private FqnResolver fqnResolver;
+  private MongoBackendResolver backendResolver;
+  private StructuredTraceGraph structuredTraceGraph;
+
+  @BeforeEach
+  public void setup() {
+    this.fqnResolver = new HypertraceFqnResolver();
+    backendResolver = new MongoBackendResolver(this.fqnResolver);
+    structuredTraceGraph = mock(StructuredTraceGraph.class);
+  }
+
+  @Test
+  public void checkBackendEntityGeneratedFromUninstrumentedMongoEvent() {
+    Event e =
+        Event.newBuilder()
+            .setCustomerId("__default")
+            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f8".getBytes()))
+            .setEntityIdList(Arrays.asList("4bfca8f7-4974-36a4-9385-dd76bf5c8824"))
+            .setEnrichedAttributes(
+                Attributes.newBuilder()
+                    .setAttributeMap(
+                        Map.of("SPAN_TYPE", AttributeValue.newBuilder().setValue("EXIT").build()))
+                    .build())
+            .setAttributes(
+                Attributes.newBuilder()
+                    .setAttributeMap(
+                        Map.of(
+                            "NAMESPACE",
+                            AttributeValue.newBuilder().setValue("sampleshop.userReview").build(),
+                            "span.kind",
+                            AttributeValue.newBuilder().setValue("client").build(),
+                            "OPERATION",
+                            AttributeValue.newBuilder().setValue("FindOperation").build(),
+                            "k8s.pod_id",
+                            AttributeValue.newBuilder()
+                                .setValue("55636196-c840-11e9-a417-42010a8a0064")
+                                .build(),
+                            "docker.container_id",
+                            AttributeValue.newBuilder()
+                                .setValue(
+                                    "ee85cf2cfc3b24613a3da411fdbd2f3eabbe729a5c86c5262971c8d8c29dad0f")
+                                .build(),
+                            "FLAGS",
+                            AttributeValue.newBuilder().setValue("0").build(),
+                            "address",
+                            AttributeValue.newBuilder().setValue(MONGO_URL).build()))
+                    .build())
+            .setEventName("mongo.async.exit")
+            .setStartTimeMillis(1566869077746L)
+            .setEndTimeMillis(1566869077750L)
+            .setMetrics(
+                Metrics.newBuilder()
+                    .setMetricMap(
+                        Map.of("Duration", MetricValue.newBuilder().setValue(4.0).build()))
+                    .build())
+            .setEventRefList(
+                Arrays.asList(
+                    EventRef.newBuilder()
+                        .setTraceId(ByteBuffer.wrap("random_trace_id".getBytes()))
+                        .setEventId(ByteBuffer.wrap("random_event_id".getBytes()))
+                        .setRefType(EventRefType.CHILD_OF)
+                        .build()))
+            .build();
+    final Entity backendEntity = backendResolver.resolve(e, structuredTraceGraph).get().getEntity();
+    assertEquals("mongo:27017", backendEntity.getEntityName());
+    assertEquals(3, backendEntity.getIdentifyingAttributesCount());
+    assertEquals(
+        backendEntity
+            .getIdentifyingAttributesMap()
+            .get(Constants.getEntityConstant(BackendAttribute.BACKEND_ATTRIBUTE_PROTOCOL))
+            .getValue()
+            .getString(),
+        "MONGO");
+    assertEquals(
+        "mongo",
+        backendEntity
+            .getIdentifyingAttributesMap()
+            .get(Constants.getEntityConstant(BackendAttribute.BACKEND_ATTRIBUTE_HOST))
+            .getValue()
+            .getString());
+    assertEquals(
+        backendEntity
+            .getIdentifyingAttributesMap()
+            .get(Constants.getEntityConstant(BackendAttribute.BACKEND_ATTRIBUTE_PORT))
+            .getValue()
+            .getString(),
+        "27017");
+    assertEquals(
+        backendEntity.getAttributesMap().get("NAMESPACE").getValue().getString(),
+        "sampleshop.userReview");
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(Constants.getEnrichedSpanConstant(Backend.BACKEND_FROM_EVENT))
+            .getValue()
+            .getString(),
+        "mongo.async.exit");
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(Constants.getEnrichedSpanConstant(Backend.BACKEND_FROM_EVENT_ID))
+            .getValue()
+            .getString(),
+        "62646630336466616266356337306638");
+  }
+
+  @Test
+  public void checkBackendEntityGeneratedFromInstrumentedMongoEvent() {
+    Event e =
+        Event.newBuilder()
+            .setCustomerId("__default")
+            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f8".getBytes()))
+            .setEntityIdList(Arrays.asList("4bfca8f7-4974-36a4-9385-dd76bf5c8824"))
+            .setEnrichedAttributes(
+                Attributes.newBuilder()
+                    .setAttributeMap(
+                        Map.of("SPAN_TYPE", AttributeValue.newBuilder().setValue("EXIT").build()))
+                    .build())
+            .setAttributes(
+                Attributes.newBuilder()
+                    .setAttributeMap(
+                        Map.of(
+                            "mongo.namespace",
+                            AttributeValue.newBuilder().setValue("sampleshop.userReview").build(),
+                            "span.kind",
+                            AttributeValue.newBuilder().setValue("client").build(),
+                            "OPERATION",
+                            AttributeValue.newBuilder().setValue("FindOperation").build(),
+                            "k8s.pod_id",
+                            AttributeValue.newBuilder()
+                                .setValue("55636196-c840-11e9-a417-42010a8a0064")
+                                .build(),
+                            "docker.container_id",
+                            AttributeValue.newBuilder()
+                                .setValue(
+                                    "ee85cf2cfc3b24613a3da411fdbd2f3eabbe729a5c86c5262971c8d8c29dad0f")
+                                .build(),
+                            "FLAGS",
+                            AttributeValue.newBuilder().setValue("0").build(),
+                            "mongo.operation",
+                            AttributeValue.newBuilder().setValue("HelloWorld").build(),
+                            "mongo.url",
+                            AttributeValue.newBuilder().setValue(MONGO_URL).build()))
+                    .build())
+            .setEventName("mongo.async.exit")
+            .setStartTimeMillis(1566869077746L)
+            .setEndTimeMillis(1566869077750L)
+            .setMetrics(
+                Metrics.newBuilder()
+                    .setMetricMap(
+                        Map.of("Duration", MetricValue.newBuilder().setValue(4.0).build()))
+                    .build())
+            .setEventRefList(
+                Arrays.asList(
+                    EventRef.newBuilder()
+                        .setTraceId(ByteBuffer.wrap("random_trace_id".getBytes()))
+                        .setEventId(ByteBuffer.wrap("random_event_id".getBytes()))
+                        .setRefType(EventRefType.CHILD_OF)
+                        .build()))
+            .build();
+    final Entity backendEntity = backendResolver.resolve(e, structuredTraceGraph).get().getEntity();
+    assertEquals("mongo:27017", backendEntity.getEntityName());
+    assertEquals(3, backendEntity.getIdentifyingAttributesCount());
+    assertEquals(
+        backendEntity
+            .getIdentifyingAttributesMap()
+            .get(Constants.getEntityConstant(BackendAttribute.BACKEND_ATTRIBUTE_PROTOCOL))
+            .getValue()
+            .getString(),
+        "MONGO");
+    assertEquals(
+        "mongo",
+        backendEntity
+            .getIdentifyingAttributesMap()
+            .get(Constants.getEntityConstant(BackendAttribute.BACKEND_ATTRIBUTE_HOST))
+            .getValue()
+            .getString());
+    assertEquals(
+        backendEntity
+            .getIdentifyingAttributesMap()
+            .get(Constants.getEntityConstant(BackendAttribute.BACKEND_ATTRIBUTE_PORT))
+            .getValue()
+            .getString(),
+        "27017");
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(Constants.getRawSpanConstant(Mongo.MONGO_NAMESPACE))
+            .getValue()
+            .getString(),
+        "sampleshop.userReview");
+
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(Constants.getRawSpanConstant(Mongo.MONGO_OPERATION))
+            .getValue()
+            .getString(),
+        "HelloWorld");
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(Constants.getEnrichedSpanConstant(Backend.BACKEND_FROM_EVENT))
+            .getValue()
+            .getString(),
+        "mongo.async.exit");
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(Constants.getEnrichedSpanConstant(Backend.BACKEND_FROM_EVENT_ID))
+            .getValue()
+            .getString(),
+        "62646630336466616266356337306638");
+  }
+
+  @Test
+  public void checkBackendEntityGeneratedFromInstrumentedMongoEventOtelFormat() {
+    Event e =
+        Event.newBuilder()
+            .setCustomerId("__default")
+            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f8".getBytes()))
+            .setEntityIdList(Arrays.asList("4bfca8f7-4974-36a4-9385-dd76bf5c8824"))
+            .setEnrichedAttributes(
+                Attributes.newBuilder()
+                    .setAttributeMap(
+                        Map.of("SPAN_TYPE", AttributeValue.newBuilder().setValue("EXIT").build()))
+                    .build())
+            .setAttributes(
+                Attributes.newBuilder()
+                    .setAttributeMap(
+                        Map.of(
+                            OTelDbSemanticConventions.DB_SYSTEM.getValue(),
+                            buildAttributeValue(
+                                OTelDbSemanticConventions.MONGODB_DB_SYSTEM_VALUE.getValue()),
+                            OTelSpanSemanticConventions.NET_PEER_NAME.getValue(),
+                            buildAttributeValue("mongodb0"),
+                            OTelDbSemanticConventions.MONGODB_COLLECTION.getValue(),
+                            buildAttributeValue("sampleshop.userReview"),
+                            "span.kind",
+                            buildAttributeValue("client"),
+                            OTelDbSemanticConventions.DB_OPERATION.getValue(),
+                            buildAttributeValue("FindOperation"),
+                            OTelDbSemanticConventions.DB_NAME.getValue(),
+                            buildAttributeValue("customers"),
+                            OTelSpanSemanticConventions.NET_PEER_PORT.getValue(),
+                            buildAttributeValue("27017")))
+                    .build())
+            .setEventName("mongo.async.exit")
+            .setStartTimeMillis(1566869077746L)
+            .setEndTimeMillis(1566869077750L)
+            .setMetrics(
+                Metrics.newBuilder()
+                    .setMetricMap(
+                        Map.of("Duration", MetricValue.newBuilder().setValue(4.0).build()))
+                    .build())
+            .setEventRefList(
+                Arrays.asList(
+                    EventRef.newBuilder()
+                        .setTraceId(ByteBuffer.wrap("random_trace_id".getBytes()))
+                        .setEventId(ByteBuffer.wrap("random_event_id".getBytes()))
+                        .setRefType(EventRefType.CHILD_OF)
+                        .build()))
+            .build();
+    final BackendInfo backendInfo = backendResolver.resolve(e, structuredTraceGraph).get();
+    final Entity backendEntity = backendInfo.getEntity();
+    assertEquals("mongodb0:27017", backendEntity.getEntityName());
+    assertEquals(3, backendEntity.getIdentifyingAttributesCount());
+    assertEquals(
+        backendEntity
+            .getIdentifyingAttributesMap()
+            .get(Constants.getEntityConstant(BackendAttribute.BACKEND_ATTRIBUTE_PROTOCOL))
+            .getValue()
+            .getString(),
+        "MONGO");
+    assertEquals(
+        "mongodb0",
+        backendEntity
+            .getIdentifyingAttributesMap()
+            .get(Constants.getEntityConstant(BackendAttribute.BACKEND_ATTRIBUTE_HOST))
+            .getValue()
+            .getString());
+    assertEquals(
+        backendEntity
+            .getIdentifyingAttributesMap()
+            .get(Constants.getEntityConstant(BackendAttribute.BACKEND_ATTRIBUTE_PORT))
+            .getValue()
+            .getString(),
+        "27017");
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(OTelDbSemanticConventions.MONGODB_COLLECTION.getValue())
+            .getValue()
+            .getString(),
+        "sampleshop.userReview");
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(OTelDbSemanticConventions.DB_OPERATION.getValue())
+            .getValue()
+            .getString(),
+        "FindOperation");
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(Constants.getEnrichedSpanConstant(Backend.BACKEND_FROM_EVENT))
+            .getValue()
+            .getString(),
+        "mongo.async.exit");
+    assertEquals(
+        backendEntity
+            .getAttributesMap()
+            .get(Constants.getEnrichedSpanConstant(Backend.BACKEND_FROM_EVENT_ID))
+            .getValue()
+            .getString(),
+        "62646630336466616266356337306638");
+
+    Map<String, AttributeValue> attributes = backendInfo.getAttributes();
+    assertEquals(
+        Map.of(
+            "BACKEND_DESTINATION",
+            AttributeValueCreator.create("customers.sampleshop.userReview"),
+            "BACKEND_OPERATION",
+            AttributeValueCreator.create("FindOperation")),
+        attributes);
+  }
+}
